Move the drink requests and state to main page
