+++ src/App.js
@@ -3,27 +3,17 @@ import RecipesPage from './Components/RecipesPage/RecipesPage'
 import './App.css'
 import { Route } from 'react-router-dom'
 import Header from './Components/Header/Header'
-import { getAllRecipes } from './APIRequests'
+import Footer from './Components/Footer/Footer'
 import DrinkDetailsPage from './Components/DrinkDetailsPage/DrinkDetailsPage'
 
 class App extends Component {
   constructor() {
     super()
     this.state = {
-      allDrinks: [],
       error: '',
       drinkId: 0
     }
     this.getDrinkDetails = this.getDrinkDetails.bind(this)
-  }
-
-  async componentDidMount() {
-    try {
-      const recipes = await getAllRecipes()
-      this.setState({ allDrinks: recipes.drinks })
-    } catch (error) {
-      this.setState({ error: error })
-    }
   }
 
   getDrinkDetails(id) {
@@ -32,14 +22,14 @@ class App extends Component {
 
   render() {
     return (
+      <>
+      <Header />
       <main>
-        <Header />
         <Route
           exact path = "/"
           render={() => {
             return (
               <RecipesPage
-                allDrinks={this.state.allDrinks}
                 getDrinkDetails={this.getDrinkDetails}
               />
             )
@@ -55,8 +45,9 @@ class App extends Component {
             )
           }}
         />
-        
       </main>
+      <Footer />
+      </>
     )
   }
 }
